Fix  : ThemeContext
